requirements.txt
datasets==4.2.0
fastapi==0.121.2
numpy==2.3.5
pandas==2.3.3
peft==0.17.1
pydantic==2.12.4
PyYAML==6.0.3
sentence_transformers==5.1.1
streamlit==1.51.0
torch==2.9.0
transformers==4.57.1
accelerate
bitsandbytes
huggingface_hub
faiss-cpu   # optional; faster. If installs fail use scikit-learn fallback (already supported)
uvicorn
scikit-learn
sacrebleu
evaluate
pandas
scipy
